+++ packages/pob-dependencies/lib/index.d.ts
@@ -1,37 +1,37 @@
 
 interface Dependencies {
-  '@babel/core': '7.5.5';
-  '@babel/preset-env': '7.5.5';
+  '@babel/core': '7.6.0';
+  '@babel/preset-env': '7.6.0';
   '@babel/preset-react': '7.0.0';
-  '@babel/runtime': '7.5.5';
+  '@babel/runtime': '7.6.0';
   '@pob/eslint-config': '32.0.2';
-  '@pob/eslint-config-babel': '32.0.2';
-  '@pob/eslint-config-babel-node': '32.0.0';
-  '@pob/eslint-config-node': '32.0.0';
-  '@pob/eslint-config-react': '32.0.0';
-  '@pob/eslint-config-typescript': '32.0.2';
-  '@pob/eslint-config-typescript-react': '32.0.0';
-  '@pob/repo-config': '^1.0.0';
-  '@pob/version': '^4.33.0';
+  '@pob/eslint-config-babel': '32.0.3';
+  '@pob/eslint-config-babel-node': '32.0.13';
+  '@pob/eslint-config-node': '32.0.13';
+  '@pob/eslint-config-react': '32.0.10';
+  '@pob/eslint-config-typescript': '32.0.13';
+  '@pob/eslint-config-typescript-react': '32.0.10';
+  '@pob/repo-config': '^1.0.1';
+  '@pob/version': '^4.33.1';
   '@types/jest': '24.0.18';
-  '@typescript-eslint/eslint-plugin': '2.0.0';
-  '@typescript-eslint/parser': '2.0.0';
+  '@typescript-eslint/eslint-plugin': '2.2.0';
+  '@typescript-eslint/parser': '2.2.0';
   'babel-eslint': '10.0.3';
   'babel-jest': '24.9.0';
   'babel-preset-latest-node': '3.2.1';
   'babel-preset-modern-browsers': '14.0.0';
-  'eslint': '6.2.2';
-  'husky': '3.0.4';
+  'eslint': '6.3.0';
+  'husky': '3.0.5';
   'jest': '24.9.0';
   'jest-junit-reporter': '1.1.0';
   'lerna': '3.16.4';
-  'pob-babel': '^23.19.1';
-  'pob-lcov-reporter': '^1.2.1';
-  'pob-release': '^4.13.3';
+  'pob-babel': '^23.20.0';
+  'pob-lcov-reporter': '^1.2.2';
+  'pob-release': '^4.14.0';
   'prettier': '1.18.2';
   'repository-check-dirty': '^1.0.2';
   'typedoc': '0.15.0';
-  'typescript': '3.6.2';
+  'typescript': '3.6.3';
 } 
 
 export = dependencies;