+++ @pob/repo-config/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/christophehurpeau/pob/compare/@pob/repo-config@1.0.0...@pob/repo-config@1.0.2) (2019-09-13)
+
+
+### Bug Fixes
+
+* **repo-config:** add yarn.lock ([3d0efd7](https://github.com/christophehurpeau/pob/commit/3d0efd7))
+
+
+
+
+
 ## [1.0.1](https://github.com/christophehurpeau/pob/compare/@pob/repo-config@1.0.0...@pob/repo-config@1.0.1) (2019-08-30)